Use structopt to implement `set input-mode emacs`
Before this commit, all backslash commands (like `\help`) could only be
exact string matches. We need a path forward to accept args, and have
decent help.

This commit takes the stance that we can use structopt (which we already
use for CLI parsing) to also parse backslash commands. The first one
I've implemented is for changing the input mode from emacs to vi.

Nothing actually happens yet other than a messing being printed to the
screen. That's because we need to actually expose the mode-change on the
Ui's editor.